refactor: participant context 생성
