+++ frontend/src/App.js
@@ -4,13 +4,12 @@ import { QueryClient, QueryClientProvider } from 'react-query';
 import { ReactQueryDevtools } from 'react-query/devtools';
 
 import { HomePage, LoginPage, LogoutPage, MidpointPage, OAuthPage } from './pages';
-import { useParticipants } from './hooks';
 import { NavBar } from './components';
+import { ParticipantContextProvider } from './contexts';
 import { ROUTE, REACT_QUERY_DEV_TOOL } from './constants';
 
 export const App = () => {
   const queryClient = new QueryClient();
-  const { participant } = useParticipants();
   const [user, setUser] = useState({});
 
   return (
@@ -18,14 +17,15 @@ export const App = () => {
       <Router>
         <NavBar user={user} />
         <Switch>
-          <Route exact path={ROUTE.HOME.PATH}>
-            <HomePage participant={participant} />
-          </Route>
-          // TODO: 중간지점 찾기 페이지
-          <Route exact path={ROUTE.MIDPOINT.PATH}>
-            <MidpointPage participant={participant} />
-          </Route>
-          // TODO: 로그인 / 로그아웃 페이지
+          <ParticipantContextProvider>
+            <Route exact path={ROUTE.HOME.PATH}>
+              <HomePage />
+            </Route>
+            <Route exact path={ROUTE.MIDPOINT.PATH}>
+              <MidpointPage />
+            </Route>
+          </ParticipantContextProvider>
+
           <Route exact path={ROUTE.LOGIN.PATH}>
             <LoginPage />
           </Route>
