+++ frontend/src/contexts/index.js
@@ -1,0 +1,1 @@
+export { ParticipantContext, ParticipantContextProvider } from './ParticipantContext';
