+++ frontend/src/hooks/index.js
@@ -3,4 +3,3 @@ export { useMapView } from './useMapView';
 export { useModal } from './useModal';
 export { useParticipantForm } from './useParticipantForm';
 export { useParticipantRemoveConfirm } from './useParticipantRemoveConfirm';
-export { useParticipants } from './useParticipants';
